+++ src/routes/index.js
@@ -2,7 +2,7 @@ const express = require('express');
 const UserRouter = require('./user.routes');
 const ProfileRouter = require('./profile.routes');
 const CategoryRouter = require('./categories.routes');
-const validateToken = require('../middlewares/authToken');
+const validateToken = require('../middlewares/auth/authToken');
 const BudgetRoutes = require('./budget.routes');
 
 const router = express.Router();
